Requisito 6

+++ routes/talkerRoutes.js
@@ -1,7 +1,7 @@
 const router = require('express').Router();
 const fs = require('fs').promises;
 
-const { readTalkerFile } = require('../helpers/validations');
+const { readTalkerFile, getTalker } = require('../helpers/validations');
 const { validateToken } = require('../helpers/validateToken');
 
 router.get('/', async (_req, res) => {
@@ -12,55 +12,58 @@ router.get('/', async (_req, res) => {
     .catch(({ message }) => res.status(500).json({ message }));
 });
 
-router.get('/:id', (req, res) => {
+router.get('/:id', async (req, res) => {
   const { id } = req.params;
-
-  readTalkerFile()
-  .then((data) => {
-    const talker = data.find((t) => Number(t.id) === Number(id));
-    if (!talker) return res.status(404).json({ message: 'Pessoa palestrante não encontrada' });
-    res.status(200).json(talker);
-  })
-  .catch(({ message }) => res.status(500).json({ message }));
+  const talker = await getTalker(id);
+  if (!talker) return res.status(404).json({ message: 'Pessoa palestrante não encontrada' });
+  res.status(200).json(talker);
 });
 
 const validateTalker = require('../helpers/validateTalker');
 
 router.use(validateToken);
-router.use(validateTalker);
 
-router.post('/', (req, res) => {
-  const { name, age, talk } = req.body;
-  const talker = { name, age, talk };
-  readTalkerFile()
-  .then(async (data) => {
-    const talkerData = {
-      ...talker,
-      id: data.length + 1,
-    };
-    await fs.writeFile('./talker.json', JSON.stringify([...data, talkerData], null, 2));
-    res.status(201).json(talkerData);
-  })
-  .catch(({ message }) => res.status(500).json({ message }));
+router.delete('/:id', async (req, res) => {
+  const { id } = req.params;
+  const exists = await getTalker(id);
+  if (!exists) return res.status(404).json({ message: 'Pessoa palestrante não encontrada' });
+
+  const data = await readTalkerFile();
+  const newData = data.filter((t) => Number(t.id) !== Number(id));
+  await fs.writeFile('./talker.json', JSON.stringify(newData, null, 2));
+  res.status(200).json({ message: 'Pessoa palestrante deletada com sucesso' });
 });
 
-router.put('/:id', (req, res) => {
+router.use(validateTalker);
+
+router.post('/', async (req, res) => {
+  const { name, age, talk } = req.body;
+  const talker = { name, age, talk };
+  const data = await readTalkerFile();
+  const talkerData = {
+    ...talker,
+    id: data.length + 1,
+  };
+  await fs.writeFile('./talker.json', JSON.stringify([...data, talkerData], null, 2));
+  res.status(201).json(talkerData);
+});
+
+router.put('/:id', async (req, res) => {
   const { id } = req.params;
   const { name, age, talk } = req.body;
+  const exists = await getTalker(id);
+  if (!exists) return res.status(404).json({ message: 'Pessoa palestrante não encontrada' });
   let talker = { name, age, talk };
-  readTalkerFile()
-    .then(async (data) => {
-      const newData = data.map((t) => {
-        if (Number(t.id) === Number(id)) {
-          talker = { ...talker, id: t.id };
-          return talker;
-        }
-        return t;
-      });
-      await fs.writeFile('./talker.json', JSON.stringify(newData, null, 2));
-      res.status(200).json(talker);
-    })
-    .catch(({ message }) => res.status(500).json({ message }));
+  const data = await readTalkerFile();
+  const newData = data.map((t) => {
+    if (Number(t.id) === Number(id)) {
+      talker = { ...talker, id: t.id };
+      return talker;
+    }
+    return t;
+  });
+  await fs.writeFile('./talker.json', JSON.stringify(newData, null, 2));
+  res.status(200).json(talker);
 });
 
 module.exports = router;
